♻️ 리팩토링 : 에러 핸들링, 타입 개선
