--- src/common.ts
@@ -1,3 +1,0 @@
-export const getRandomNumber = (range: number) => {
-  return Math.floor(Math.random() * range);
-};
