+++ src/util.ts
@@ -1,5 +1,7 @@
-import { getRandomNumber } from './common';
-
 export const getRandomItem = (inventory: string[]) => {
   return inventory[getRandomNumber(inventory.length)];
+};
+
+export const getRandomNumber = (range: number) => {
+  return Math.floor(Math.random() * range);
 };
